fix: generate key on startup of php

+++ .docker/php/entrypoint.sh
@@ -4,10 +4,6 @@ sudo find ./bootstrap/cache ./storage ! -path */\.gitignore -type f -exec chmod 
 sudo find ./bootstrap/cache ./storage -type d -exec chmod 775 {} \; -exec chown php:www-data {} \;
 
 # Setup composer dependencies
-if [ ! -f ".env" ]; then
-    cp -f .env.example .env
-fi
-
 if [ ! -d "vendor" ]; then
     composer install
 fi
@@ -16,7 +12,11 @@ fi
 echo "PostgreSQL isn't available - Waiting for it.."
 until pg_isready --host=db --port=5432 --dbname=ngts_iue_dmd --username=root &> /dev/null; do sleep 1; done
 
-# Run migrations and seeds
-php artisan migrate:refresh --seed
+# Setup env files and migrations
+if [ ! -f ".env" ]; then
+    cp -f .env.example .env
+    php artisan key:generate
+    php artisan migrate:refresh --seed
+fi
 
 exec sudo -E "${@}"
